pyproject.toml
[project]
name = "catprint"
version = "0.1.0"
description = "Print images on the MX06 (C15) thermal printer"
readme = "README.md"
authors = [
    { name = "Jan Pokorný", email = "jenompokorny@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "bleak>=1.0.1",
    "crc8>=0.2.1",
    "pdf2image>=1.17.0",
    "pillow>=11.3.0",
    "streamlit>=1.48.0",
    "uvicorn>=0.38.0",
]

[build-system]
requires = ["uv_build>=0.7.13,<0.8"]
build-backend = "uv_build"
